feat: block unauthorized author of article

+++ src/main/java/techcourse/myblog/controller/ArticleController.java
@@ -4,10 +4,17 @@ import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import org.springframework.web.servlet.view.RedirectView;
 import techcourse.myblog.domain.Article;
+import techcourse.myblog.domain.User;
 import techcourse.myblog.exception.NotFoundArticleException;
+import techcourse.myblog.exception.UnauthenticatedUserException;
 import techcourse.myblog.repository.ArticleRepository;
 import techcourse.myblog.service.dto.ArticleDTO;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 @Controller
 @RequestMapping("/articles")
@@ -19,31 +26,57 @@ public class ArticleController {
     }
 
     @GetMapping("/{articleId}")
-    public String showArticle(@PathVariable("articleId") Long articleId, Model model) {
+    public String showArticle(@PathVariable("articleId") Long articleId, HttpSession session, Model model) {
         Article article = articleRepository.findById(articleId).orElseThrow(() ->
                 new NotFoundArticleException("게시물이 없습니다"));
 
         model.addAttribute("article", article);
         model.addAttribute("comments", article.getComments());
-//        model.addAttribute("articleId", articleId);
+
+        User user = (User) session.getAttribute("user");
+        if (user != null) {
+            model.addAttribute("user", user);
+        }
 
         return "article";
     }
 
     @GetMapping("/new")
-    public String articleCreateForm() {
+    public String articleCreateForm(HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) {
+            throw new UnauthenticatedUserException("로그인이 필요합니다.");
+        }
         return "article-edit";
     }
 
     @PostMapping("/write")
-    public String saveArticle(ArticleDTO articleDTO) {
-        Article article = articleRepository.save(articleDTO.toDomain());
+    public String saveArticle(ArticleDTO articleDTO, HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) {
+            throw new UnauthenticatedUserException("로그인이 필요합니다.");
+        }
+
+        Article article = articleDTO.toDomain();
+        article.setAuthor(user);
+
+        articleRepository.save(article);
         return "redirect:/articles/" + article.getId();
     }
 
     @GetMapping("/{articleId}/edit")
-    public String articleUpdateForm(@PathVariable("articleId") Long articleId, Model model) {
-        model.addAttribute("article", articleRepository.findById(articleId).get());
+    public Object articleUpdateForm(@PathVariable("articleId") Long articleId, HttpSession session, Model model) {
+        Article savedArticle = articleRepository
+                .findById(articleId)
+                .orElseThrow(() ->
+                        new NotFoundArticleException("게시글이 존재하지 않습니다."));
+
+        User user = (User) session.getAttribute("user");
+        if (user != savedArticle.getAuthor()) {
+            throw new UnauthenticatedUserException("수정 권한이 없습니다.");
+        }
+
+        model.addAttribute("article", savedArticle);
         model.addAttribute("articleId", articleId);
         return "article-edit";
     }
@@ -58,8 +91,24 @@ public class ArticleController {
     }
 
     @DeleteMapping("/{articleId}")
-    public String deleteArticle(@PathVariable("articleId") Long articleId) {
+    public String deleteArticle(@PathVariable("articleId") Long articleId, HttpSession session) {
+        Article savedArticle = articleRepository
+                .findById(articleId)
+                .orElseThrow(() ->
+                        new NotFoundArticleException("게시글이 존재하지 않습니다."));
+
+        User user = (User) session.getAttribute("user");
+        if (user != savedArticle.getAuthor()) {
+            throw new UnauthenticatedUserException("삭제 권한이 없습니다.");
+        }
+
         articleRepository.deleteById(articleId);
         return "redirect:/";
+    }
+
+    @ExceptionHandler(RuntimeException.class)
+    public RedirectView articleExceptionHandler(RuntimeException e, RedirectAttributes redirectAttributes, HttpServletRequest request) {
+        redirectAttributes.addFlashAttribute("articleError", e.getMessage());
+        return new RedirectView(request.getHeader("Referer"));
     }
 }
